requirements.txt
pillow==11.2.1
python-telegram-bot==22.0
